+++ src/in-house-bot/commands/listmyversions.ts
@@ -1,3 +1,4 @@
+import pick from 'lodash/pick'
 import { ICommand } from '../types'
 
 export const command:ICommand = {
@@ -5,9 +6,10 @@ export const command:ICommand = {
   name: 'listmyversions',
   examples: [
     '/listmyversions',
+    '/listmyversions --limit 1',
   ],
   description: 'list previous versions of this MyCloud',
   exec: async ({ ctx, commander, req, args }) => {
-    return await commander.deployment.listMyVersions()
+    return await commander.deployment.listMyVersions(pick(args, ['limit']))
   }
 }