+++ src/in-house-bot/index.ts
@@ -320,8 +320,8 @@ export const loadComponentsAndPlugins = ({
   if (runAsyncHandlers) {
     logger.debug('running async hooks')
     // productsAPI.removeDefaultHandlers()
-    const processChange = async ({ old, value }: ISaveEventPayload) => {
-      const type = old[TYPE]
+    const maybeNotifyChildDeploymentCreators = ({ old={}, value={} }: ISaveEventPayload) => {
+      const type = value[TYPE]
       if (type === CHILD_DEPLOYMENT && didPropChange({ old, value, prop: 'stackId' })) {
         // using bot.tasks is hacky, but because this fn currently purposely stalls for minutes on end,
         // stream-processor will time out processing this item and the lambda will exit before anyone gets notified
@@ -332,7 +332,12 @@ export const loadComponentsAndPlugins = ({
 
         return
       }
+    }
 
+    const processChange = async ({ old, value }: ISaveEventPayload) => {
+      maybeNotifyChildDeploymentCreators({ old, value })
+
+      const type = old[TYPE]
       if (type === APPLICATION &&
         didPropChangeTo({ old, value, prop: 'status', propValue: 'approved' })) {
         value.submissions = await bot.backlinks.getBacklink({
@@ -348,10 +353,12 @@ export const loadComponentsAndPlugins = ({
     }
 
     const processCreate = async (resource: ITradleObject) => {
+      maybeNotifyChildDeploymentCreators({ old: null, value: resource })
+
       const type = resource[TYPE]
       if (type === VERSION_INFO &&
         resource._org === TRADLE.PERMALINK &&
-        Deployment.isMainlineReleaseTag(resource.tag)) {
+        Deployment.isStableReleaseTag(resource.tag)) {
         await alerts.updateAvailable({
           current: bot.version,
           update: resource as VersionInfo
@@ -757,7 +764,7 @@ const banter = (components: IBotComponents) => {
       application.relationshipManagers.length) return
 
     const lowercase = message.toLowerCase()
-    if (/^hey|hi|hello$/.test(message)) {
+    if (/^hey|hi|hello$/i.test(message)) {
       await productsAPI.send({
         req,
         to: user,